Fix OpenFaaS client authentication to use authentication from Terraform provider initialization
Currently, this provider was accepting `user_name` and `password` in Terraform provider initialization, but there is no reference in the code that uses it to instantiate the OpenFaaS client.

This commit adds a credentials chain so it tries to get the authentication in the following order:

1. Authentication with basic-auth using username and password from Terraform provider initialization
2. Authentication with basic-auth using OpenFaaS config file created from `faas-cli login`
3. Authentication with token using OpenFaaS config file created from `faas-cli login`

Also, adds a few log messages to help understand which credential is choosen from the chain.